Move transit headers to include/.
Resolves #62.
