+++ transit/include/eclipse.h
