+++ transit/include/extinction.h
