+++ transit/include/opacity.h
