+++ transit/include/transitstd.h
